acpuclock-8064: merged ktoonsez's sysfs module for enabling and disabling CPU "oc mode"
This came in very handy on the s4 as it has a bug where the first few ms of the boot
it ramps up to the highest cpu freq allowed which causes some phones to crash. With
this implementation CPU OC is disabled on boot and is either enabled manually by the
user via an app (Ktweaker) or just have a script either embedded in ramdisk or on the
system partition that senables OC after a few secs of booting.